+++ common-proto/src/main/proto/systems/ajax/malov/commonmodel/stock/big_decimal.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package systems.ajax.malov.commonproto.stock;
+package systems.ajax.malov.commonmodel.stock;
 
 option java_multiple_files = true;
