+++ common-proto/src/main/proto/systems/ajax/malov/commonmodel/stock/short_stock_response.proto
@@ -1,0 +1,14 @@
+syntax = "proto3";
+
+package systems.ajax.malov.commonmodel.stock;
+
+import "systems/ajax/malov/commonmodel/stock/big_decimal.proto";
+
+option java_multiple_files = true;
+
+message ShortStockRecord{
+  commonmodel.stock.BigDecimalProto open_price = 1;
+  commonmodel.stock.BigDecimalProto high_price = 2;
+  commonmodel.stock.BigDecimalProto low_price = 3;
+  commonmodel.stock.BigDecimalProto current_price = 4;
+}
