+++ common-proto/src/main/proto/systems/ajax/malov/commonmodel/stock/stock_price.proto
@@ -1,14 +1,14 @@
 syntax = "proto3";
 
-import "systems/ajax/malov/commonproto/stock/big_decimal.proto";
+import "systems/ajax/malov/commonmodel/stock/big_decimal.proto";
 import "google/protobuf/timestamp.proto";
 
-package systems.ajax.malov.commonproto.stock;
+package systems.ajax.malov.commonmodel.stock;
 
 option java_multiple_files = true;
 
 message StockPrice {
   string stock_symbol_name = 1;
-  commonproto.stock.BigDecimalProto price = 2;
+  commonmodel.stock.BigDecimalProto price = 2;
   google.protobuf.Timestamp timestamp = 3;
 }
