--- common-proto/src/main/proto/systems/ajax/malov/commonproto/stock/short_stock_response.proto
@@ -1,14 +1,0 @@
-syntax = "proto3";
-
-package systems.ajax.malov.commonproto.stock;
-
-import "systems/ajax/malov/commonproto/stock/big_decimal.proto";
-
-option java_multiple_files = true;
-
-message ShortStockRecord{
-  commonproto.stock.BigDecimalProto open_price = 1;
-  commonproto.stock.BigDecimalProto high_price = 2;
-  commonproto.stock.BigDecimalProto low_price = 3;
-  commonproto.stock.BigDecimalProto current_price = 4;
-}
