+++ domainservice/src/main/kotlin/systems/ajax/malov/stockanalyzer/dto/NotificationStockPriceDto.kt
@@ -1,7 +1,7 @@
 package systems.ajax.malov.stockanalyzer.dto
 
 import org.bson.types.ObjectId
-import systems.ajax.malov.commonproto.stock.StockPrice
+import systems.ajax.malov.commonmodel.stock.StockPrice
 
 data class NotificationStockPriceDto(
     val stockPrice: StockPrice,
