+++ domainservice/src/main/kotlin/systems/ajax/malov/stockanalyzer/service/NotificationStockPriceService.kt
@@ -4,7 +4,7 @@ import org.bson.types.ObjectId
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import reactor.util.function.Tuple2
-import systems.ajax.malov.commonproto.stock.StockPrice
+import systems.ajax.malov.commonmodel.stock.StockPrice
 import systems.ajax.malov.internalapi.output.pubsub.stock.NotificationStockPrice
 
 interface NotificationStockPriceService {
