+++ internal-api/src/main/proto/systems/ajax/malov/internalapi/input/reqreply/stock/get_best_stock_symbols_with_records.proto
@@ -2,8 +2,8 @@ syntax = "proto3";
 
 package systems.ajax.malov.internalapi.input.reqreply.stock;
 
-import "systems/ajax/malov/commonproto/stock/big_decimal.proto";
-import "systems/ajax/malov/commonproto/stock/short_stock_response.proto";
+import "systems/ajax/malov/commonmodel/stock/big_decimal.proto";
+import "systems/ajax/malov/commonmodel/stock/short_stock_response.proto";
 
 
 option java_multiple_files = true;
@@ -29,5 +29,5 @@ message GetBestStockSymbolsWithStockRecordsResponse {
 
 message AggregatedStockRecordItemResponse{
   string stock_symbol = 1;
-  repeated commonproto.stock.ShortStockRecord data = 2;
+  repeated commonmodel.stock.ShortStockRecord data = 2;
 }
