+++ internal-api/src/main/proto/systems/ajax/malov/internalapi/output/pubsub/stock/notification_stock_price.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package systems.ajax.malov.internalapi.output.pubsub.stock;
 
-import "systems/ajax/malov/commonproto/stock/big_decimal.proto";
+import "systems/ajax/malov/commonmodel/stock/big_decimal.proto";
 import "google/protobuf/timestamp.proto";
 
 
@@ -10,7 +10,7 @@ option java_multiple_files = true;
 
 message NotificationStockPrice {
   string stock_symbol_name = 1;
-  commonproto.stock.BigDecimalProto price = 2;
+  commonmodel.stock.BigDecimalProto price = 2;
   string user_id = 3;
   google.protobuf.Timestamp timestamp = 4;
 }
